Refs #6 Train network with a single image vector

+++ src/models/mnist.h
@@ -9,6 +9,6 @@
 #include <opencv2/core/eigen.hpp>
 #include <opencv2/opencv.hpp>
 
-Eigen::VectorXf convert_image_to_vector(cv::Mat image);
+Eigen::VectorXd convert_image_to_vector(cv::Mat image);
 
 #endif //CYNAPSE_MNIST_CPP_H
